+++ test/ttMarkdown.test.ts
@@ -1,67 +1,67 @@
-import * as assert from 'assert';
-import { MarkdownParser } from '../src/ttMarkdown';
-import { RowType } from '../src/ttTable';
+import * as assert from 'assert'
+import { MarkdownParser } from '../src/ttMarkdown'
+import { RowType } from '../src/ttTable'
 
 suite('Text tables. Markdown', () => {
     suite('Parser', () => {
-        let parser: MarkdownParser;
+        let parser: MarkdownParser
         setup(() => {
-            parser = new MarkdownParser();
-        });
+            parser = new MarkdownParser()
+        })
 
         test('should return undefined when incorrect text provided', () => {
-            const table = parser.parse('');
-            assert.equal(table, undefined);
-        });
+            const table = parser.parse('')
+            assert.equal(table, undefined)
+        })
 
         test('should return table when correct text provided', () => {
-            const table = parser.parse('a');
-            assert.notEqual(table, undefined);
-        });
+            const table = parser.parse('a')
+            assert.notEqual(table, undefined)
+        })
 
         test('should add row when string starts with |', () => {
-            const table = parser.parse('|');
+            const table = parser.parse('|')
             if (table !== undefined) {
-                assert.equal(table.rows.length, 1);
-                assert.equal(table.rows[0].type, RowType.data);
-                assert.equal(table.cols.length, 1);
+                assert.equal(table.rows.length, 1)
+                assert.equal(table.rows[0].type, RowType.data)
+                assert.equal(table.cols.length, 1)
             }
-        });
+        })
 
         test('should parse "| -" as separator row', () => {
-            const table = parser.parse('| -');
-            assert.notEqual(table, undefined);
+            const table = parser.parse('| -')
+            assert.notEqual(table, undefined)
             if (table !== undefined) {
-                assert.equal(table.rows.length, 1);
-                assert.equal(table.rows[0].type, RowType.separator);
+                assert.equal(table.rows.length, 1)
+                assert.equal(table.rows[0].type, RowType.separator)
             }
-        });
+        })
 
         test('should parse "| :-" as separator row', () => {
-            const table = parser.parse('| :-');
-            assert.notEqual(table, undefined);
+            const table = parser.parse('| :-')
+            assert.notEqual(table, undefined)
             if (table !== undefined) {
-                assert.equal(table.rows.length, 1);
-                assert.equal(table.rows[0].type, RowType.separator);
+                assert.equal(table.rows.length, 1)
+                assert.equal(table.rows[0].type, RowType.separator)
             }
-        });
+        })
 
         test('should split columns by |', () => {
-            const table = parser.parse('||||');
+            const table = parser.parse('||||')
             if (table !== undefined) {
-                assert.equal(table.rows.length, 1);
-                assert.equal(table.cols.length, 3);
+                assert.equal(table.rows.length, 1)
+                assert.equal(table.cols.length, 3)
             }
-        });
+        })
 
         test('should handle last characters as column', () => {
-            const table = parser.parse('||Last col');
+            const table = parser.parse('||Last col')
             if (table !== undefined) {
-                assert.equal(table.cols.length, 2);
+                assert.equal(table.cols.length, 2)
             }
             else {
                 assert.fail('table should be parsable')
             }
-        });
-    });
-});
+        })
+    })
+})