+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# 1.3.0 - 2021-03-20
+
+### Added
+
+- Column alignments: 
+  - left `|:--- |`
+  - right `| ---:|`
+  - centered `|:---:|`
+  - default `| --- |`
+
+### Changed
+
+- Improved table prefix handling
+
 # 1.2.1 - 2021-03-14
 
 ### Changed